fix: update migrations to use bigint(20) unsigned and execute all modern features migrations
Migration files updated (6 files):
- 040_update_ai_conversations_structure.sql - New migration to fix ai_conversations table
- 041_sustainability_carbon_tracking.sql - Updated to use bigint(20) unsigned
- 042_whatsapp_integration.sql - Updated to use bigint(20) unsigned
- 043_business_operations.sql - Updated to use bigint(20) unsigned
- 044_document_trip_management.sql - Updated to use bigint(20) unsigned
- 045_social_features.sql - Updated to use bigint(20) unsigned

Migrations executed successfully:
- Module 40: ai_conversations (updated structure)
- Module 41: carbon_emissions, eco_scores, eco_actions, low_carbon_routes (4 tables)
- Module 42: whatsapp_contacts, whatsapp_messages, whatsapp_templates, whatsapp_campaigns (4 tables)
- Module 43: guide_match_results, smart_schedule_entries, payroll_records, gps_clock_in_records, express_book_records (5 tables)
- Module 44: digital_wallet, wallet_transactions, imported_itineraries, trip_timeline_entries, pdf_itineraries, websocket_subscriptions (6 tables)
- Module 45: group_trips, group_trip_participants, shared_wishlists, shared_wishlist_items, shared_wishlist_collaborators, split_payment_groups, split_payment_members, trip_albums, trip_album_photos, trip_album_comments (10 tables)

Total: 29 new tables created for modern features (Modules 40-45)
All migrations verified successfully

+++ database/migrations/041_sustainability_carbon_tracking.sql
@@ -5,9 +5,9 @@
 
 -- Carbon emissions table
 CREATE TABLE IF NOT EXISTS `carbon_emissions` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `user_id` int(11) NOT NULL,
-  `booking_id` int(11) DEFAULT NULL,
+  `id` bigint(20) unsigned NOT NULL AUTO_INCREMENT,
+  `user_id` bigint(20) unsigned NOT NULL,
+  `booking_id` bigint(20) unsigned DEFAULT NULL,
   `emission_type` enum('transport','accommodation','food','activity') NOT NULL,
   `transport_mode` varchar(50) DEFAULT NULL COMMENT 'car, bus, train, flight, etc.',
   `distance_km` decimal(10,2) DEFAULT NULL,
@@ -25,9 +25,9 @@ CREATE TABLE IF NOT EXISTS `carbon_emissions` (
 
 -- Eco scores table
 CREATE TABLE IF NOT EXISTS `eco_scores` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `user_id` int(11) NOT NULL,
-  `destination_id` int(11) DEFAULT NULL,
+  `id` bigint(20) unsigned NOT NULL AUTO_INCREMENT,
+  `user_id` bigint(20) unsigned NOT NULL,
+  `destination_id` bigint(20) unsigned DEFAULT NULL,
   `score` int(11) NOT NULL COMMENT '0-100 eco score',
   `level` enum('bronze','silver','gold','platinum') NOT NULL,
   `total_co2_saved` decimal(10,2) DEFAULT 0 COMMENT 'Total CO2 saved in kg',
@@ -44,8 +44,8 @@ CREATE TABLE IF NOT EXISTS `eco_scores` (
 
 -- Eco actions table
 CREATE TABLE IF NOT EXISTS `eco_actions` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `user_id` int(11) NOT NULL,
+  `id` bigint(20) unsigned NOT NULL AUTO_INCREMENT,
+  `user_id` bigint(20) unsigned NOT NULL,
   `action_type` enum('public_transport','eco_accommodation','local_food','carbon_offset','waste_reduction') NOT NULL,
   `description` varchar(255) NOT NULL,
   `co2_saved_kg` decimal(10,2) DEFAULT 0,
@@ -60,9 +60,9 @@ CREATE TABLE IF NOT EXISTS `eco_actions` (
 
 -- Low-carbon routes table
 CREATE TABLE IF NOT EXISTS `low_carbon_routes` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `origin_id` int(11) NOT NULL COMMENT 'Destination ID',
-  `destination_id` int(11) NOT NULL COMMENT 'Destination ID',
+  `id` bigint(20) unsigned NOT NULL AUTO_INCREMENT,
+  `origin_id` bigint(20) unsigned NOT NULL COMMENT 'Destination ID',
+  `destination_id` bigint(20) unsigned NOT NULL COMMENT 'Destination ID',
   `transport_mode` varchar(50) NOT NULL,
   `duration_hours` decimal(5,2) NOT NULL,
   `co2_kg` decimal(10,2) NOT NULL,
